[REFACTOR] #233 게시글/댓글 상태 처리 및 스타일링 공통 유틸리티 추출
